+++ awi/common/src/test/java/com/yanny/awi/test/TreeDecoratorTooltipTest.java
@@ -81,8 +81,8 @@ public class TreeDecoratorTooltipTest {
                 "          -> waterlogged: false",
                 "  -> Required Empty Blocks: 2",
                 "  -> Directions:",
-                "    -> UP",
-                "    -> DOWN"
+                "    -> Up",
+                "    -> Down"
         ));
     }
 }